Add "customizable" type (that's broken)
This adds "customizable" dropdown type; see Example6.

The idea is that instead of filtering, one starts typing.. and upon
blur/enter the typed text will be turned into a new item and that item
selected.

There are two problems with this:

- on iOS even the Dropdown.filterable (Example4) is hopelessly broken;
  thus the new (Example6) Dropdown.customizable is broken too
- on desktop using double-click to highlight the entire text is also
  broken (it is registered as two separate clicks, thus commiting the
  item)

And while these two issues could be fixed (maybe when I get more time),
I'm commiting it as is... for archival purposes.